+++ internal/store/mysqldb/characterstore.go
