+++ pkg/summit/serworm/proxy.go
@@ -6,8 +6,8 @@ import (
 	"net"
 	"os"
 
+	"github.com/paalgyula/summit/internal/store/localdb"
 	"github.com/paalgyula/summit/pkg/store"
-	"github.com/paalgyula/summit/pkg/store/localdb"
 	"github.com/paalgyula/summit/pkg/summit/auth"
 	"github.com/paalgyula/summit/pkg/summit/client"
 	"github.com/paalgyula/summit/pkg/summit/world"
